Move serializeQuery into its own file
