+++ .gitignore
@@ -4,3 +4,4 @@ vendor
 dist
 .DS_Store
 node_modules
+firestore-debug.log
